requirements.txt
aiofiles==25.1.0
aiogram==2.25.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
annotated-types==0.7.0
APScheduler==3.11.2
async-timeout==5.0.1
attrs==26.1.0
babel==2.18.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.6
cycler==0.12.1
Deprecated==1.3.1
distro==1.9.0
fonttools==4.62.1
frozenlist==1.8.0
google-api-core==2.30.1
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
googleapis-common-protos==1.73.1
httplib2==0.31.2
idna==3.11
kiwisolver==1.5.0
magic-filter==1.0.12
matplotlib==3.10.8
multidict==6.7.1
mysql-connector-python==9.6.0
netaddr==1.3.0
numpy==2.4.4
oauth2client==4.1.3
packaging==26.0
pillow==12.1.1
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyfiglet==1.0.4
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
requests==2.33.1
rsa==4.9.1
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
wrapt==2.1.2
yarl==1.23.0
yookassa==3.10.0
